skills/skill_draft/requirements.txt
# skill_draft requirements
# Python packages necesarios para el servicio

# Core
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# LLM Backend
llama-cpp-python==0.2.90

# Utilities
psutil==5.9.8
huggingface-hub==0.20.1

# Optional (para debugging)
pydantic==2.5.3
